+++ contracts/token/onft/IONFT.sol
@@ -14,14 +14,7 @@ interface IONFT is IERC721 {
      * `_zroPaymentAddress` set to address(0x0) if not paying in ZRO (LayerZero Token)
      * `_adapterParam` is a flexible bytes array to indicate messaging adapter services
      */
-    function send(
-        uint16 _dstChainId,
-        bytes calldata _toAddress,
-        uint256 _tokenId,
-        address payable _refundAddress,
-        address _zroPaymentAddress,
-        bytes calldata _adapterParam
-    ) external payable;
+    function send(uint16 _dstChainId, bytes calldata _toAddress, uint _tokenId, address payable _refundAddress, address _zroPaymentAddress, bytes calldata _adapterParam) external payable;
 
     /**
      * @dev send token `_tokenId` to (`_dstChainId`, `_toAddress`) from `_from`
@@ -29,24 +22,16 @@ interface IONFT is IERC721 {
      * `_zroPaymentAddress` set to address(0x0) if not paying in ZRO (LayerZero Token)
      * `_adapterParam` is a flexible bytes array to indicate messaging adapter services
      */
-    function sendFrom(
-        address _from,
-        uint16 _dstChainId,
-        bytes calldata _toAddress,
-        uint256 _tokenId,
-        address payable _refundAddress,
-        address _zroPaymentAddress,
-        bytes calldata _adapterParam
-    ) external payable;
+    function sendFrom(address _from, uint16 _dstChainId, bytes calldata _toAddress, uint _tokenId, address payable _refundAddress, address _zroPaymentAddress, bytes calldata _adapterParam) external payable;
 
     /**
      * @dev Emitted when `_tokenId` are moved from the `_sender` to (`_dstChainId`, `_toAddress`)
      * `_nonce` is the outbound nonce from
      */
-    event SendToChain(address indexed _sender, uint16 indexed _dstChainId, bytes indexed _toAddress, uint256 _tokenId, uint64 _nonce);
+    event SendToChain(address indexed _sender, uint16 indexed _dstChainId, bytes indexed _toAddress, uint _tokenId, uint64 _nonce);
 
     /**
      * @dev Emitted when `_tokenId` are sent from `_srcChainId` to the `_toAddress` at this chain. `_nonce` is the inbound nonce.
      */
-    event ReceiveFromChain(uint16 _srcChainId, address _toAddress, uint256 _tokenId, uint64 _nonce);
+    event ReceiveFromChain(uint16 _srcChainId, address _toAddress, uint _tokenId, uint64 _nonce);
 }
