+++ contracts/token/onft/extension/UniversalONFT.sol
@@ -6,9 +6,8 @@ import ".././ONFT.sol";
 
 /// @title Interface of the UniversalONFT standard
 contract UniversalONFT is ONFT {
-
-    uint256 startMintIndex;
-    uint256 maxMint;
+    uint startMintIndex;
+    uint maxMint;
 
     /// @notice Constructor for the UniversalONFT
     /// @param _name the name of the token
@@ -16,13 +15,7 @@ contract UniversalONFT is ONFT {
     /// @param _layerZeroEndpoint handles message transmission across chains
     /// @param _startMintIndex the starting mint number on this chain
     /// @param _maxMint the max number of mints on this chain
-    constructor(
-        string memory _name,
-        string memory _symbol,
-        address _layerZeroEndpoint,
-        uint256 _startMintIndex,
-        uint256 _maxMint
-    ) ONFT(_name, _symbol, _layerZeroEndpoint) {
+    constructor(string memory _name, string memory _symbol, address _layerZeroEndpoint, uint _startMintIndex, uint _maxMint) ONFT(_name, _symbol, _layerZeroEndpoint) {
         startMintIndex = _startMintIndex;
         maxMint = _maxMint;
     }
